requirements.txt
amqp==2.1.4
appdirs==1.4.2
arrow==0.10.0
attrs==16.3.0
Automat==0.5.0
beautifulsoup4==4.5.3
billiard==3.5.0.2
bs4==0.0.1
celery==4.0.2
cffi==1.9.1
click==6.7
constantly==15.1.0
croniter==0.3.15
cryptography==1.7.2
cssselect==1.0.1
decorator==4.0.11
fake-useragent==0.1.5
Flask==0.12
html5lib==0.999999999
idna==2.4
incremental==16.10.1
infinity==1.4
intervals==0.8.0
ipython==5.3.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.9.5
kombu==4.0.2
lxml==3.7.3
MarkupSafe==1.0
numpy==1.12.0
packaging==16.8
pandas==0.19.2
parsel==1.1.0
peewee==2.9.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.13
ptyprocess==0.5.1
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycparser==2.17
PyDispatcher==2.0.5
PyExecJS==1.4.0
Pygments==2.2.0
PyMySQL==0.7.10
pyOpenSSL==16.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2016.10
queuelib==1.4.2
redis==2.10.5
requests==2.13.0
rq==0.7.1
rq-dashboard==0.3.7
rq-scheduler==0.7.0
Scrapy==1.3.2
selenium==3.0.2
service-identity==16.0.0
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.6
SQLAlchemy-Utils==0.32.12
traitlets==4.3.2
Twisted==17.1.0
vine==1.1.3
w3lib==1.17.0
wcwidth==0.1.7
webencodings==0.5
Werkzeug==0.12.1
xlrd==1.0.0
zope.interface==4.3.3
